+++ CHANGELOG.md
@@ -1,4 +1,13 @@
 # Changelog
 
+# v0.0.2
+- Changed application name
+- Removed check for maximum game ID length, but it is now limited to int number (19 characters).
+- Added check that the input field is empty
+- Added path check
+- Added possibility to set path via browse button
+- Added state for import button, now if fields are not valid, it will not be clicked.
+- Changed jpeg quality to 100% (previously it was 80%)
+
 # v0.0.1
 - The initiation version, core functionality works, but there is no error handling or boundary condition handling.